requirements.txt
tensorflow==2.15.0
keras==2.15.0
opencv-python==4.8.1.78
mediapipe==0.10.8
streamlit==1.29.0
streamlit-webrtc==0.47.1
numpy==1.24.3
scikit-learn==1.3.2
matplotlib==3.8.2
Pillow==10.1.0
av==11.0.0